+++ PTC_xcode/PTC_xcode/city/Intersection.cpp
@@ -70,10 +70,10 @@ void Intersection::replaceRoads(GenericMap<long int, Road*>* roads)
     this->roads = roads;
 }
 
-GenericMap<int, Link*>* Intersection::getOutgoingLinks() {
-	GenericMap<int, Link*>* outGoingLinks = new GenericMap<int, Link*>();
+GenericMap<long int, Link*>* Intersection::getOutgoingLinks() {
+	GenericMap<long int, Link*>* outGoingLinks = new GenericMap<long int, Link*>();
 
-	int linkCount = 0;
+	long int linkCount = 0;
 	this->roads->initializeCounter();
 	GenericEntry<long int, Road*>* nextRoad = this->roads->nextEntry();
 	while(nextRoad != NULL)
